chore: move article type to features/articles/type for better organization
