+++ src/features/article/components/organisms/ArticleHeader.tsx
@@ -1,8 +1,8 @@
 import Breadcrumbs from '@/features/article/components/atoms/Breadcrumbs';
 import Keywords from '@/features/article/components/atoms/Keywords';
 import ArticleDetails from '../molecules/ArticleDetails';
-import { ArticleDetailsProp } from '@/types/article.types';
-import { ArticleDetailsType } from '@/types/article.types';
+import { ArticleDetailsProp } from '@/features/article/types/article.types';
+import { ArticleDetailsType } from '@/features/article/types/article.types';
 
 import { IoHeartCircleSharp } from 'react-icons/io5';
 import { ImBubble } from 'react-icons/im';
